pyproject.toml
[tool.poetry]
name = "sempregreen"
version = "0.1.0"
description = ""
authors = ["example <contact@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
django-extensions = "^3.2.1"
django-model-utils = "^4.3.1"
django-environ = "^0.9.0"
requests = "^2.28.1"
selectolax = "^0.3.11"
grequests = "^0.6.0"
django-rq = "^2.6.0"
psycopg2 = "^2.9.5"
undetected-chromedriver = "^3.1.7"
selenium-wire = "^5.1.0"
mercadopago = "^2.2.0"
gunicorn = "^20.1.0"
django-loginas = "^0.3.11"
python-dateutil = "^2.8.2"
django-admin-autocomplete-filter = "^0.7.1"
whitenoise = "^6.4.0"
django-storages = "^1.13.2"
django-bootstrap-v5 = "^1.0.11"
python-telegram-bot = "^20.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
